@@ -6,15 +6,22 @@ import App from './App';
 import * as serviceWorker from './serviceWorker';
 
 import { OpenAPI } from '@openapi/.';
-const config = require('../../config.json');
+import { ConfigurationController } from '@openapi/routes'
+import { ConfigContext } from './contexts'
 
-if (config.loopback !== undefined) {
-    OpenAPI.URL = config.loopback;
-}
-else {
-    OpenAPI.URL = window.location.origin;
-}
+// CONFIGURATION
+OpenAPI.URL = window.location.origin;
+export const Configuration = React.createContext({});
 
-ReactDOM.render(<App />, document.getElementById('root'));
+//RENDER
+ConfigurationController.getConfiguration()
+    .then(configuration => {
+        ReactDOM.render(
+            <ConfigContext.Provider value={configuration}>
+                <App />
+            </ConfigContext.Provider>,
+            document.getElementById('root')
+        );
+    })
 
 serviceWorker.unregister();
